fix: pick the correct fetcher
We need the (and) in `doc-popup-may-use-fetcher` to pick a fetcher when
both the following conditions are true:

1. The current buffer's major mode matches the 'modes property on the
fecther.
2. The predicate return t

Before this fix, we used `(or major-mode-matches predicate)` which
doesn't make sense because if predicate is nil, then it's wrapped in the
`doc-popup-define-generic-fetcher` to return t.  So, before this fix we
would always pick the first fetcher in `doc-popup-fetcher`.